+++ pollinations/ai/__init__.py
@@ -3,7 +3,9 @@ pollinations.ai
 
 Classes:
     ImageModel (ai.ImageModel): Text-to-image generative AI model.
-    TextModel (ai.TextModel): Text-to-text generative AI model.
+    (deprecated) > TextModel (ai.TextModel): Text-to-text generative AI model.
+    ImageObject (ai.ImageObject): Image object.
+    (deprecated) > TextObject (ai.TextObject): Text object.
 
 Functions:
     sample(str): Returns a sample prompt for the Image model.
@@ -28,11 +30,13 @@ from .. import abc
 from .. import ext
 from ..types import (
     ImageModel,
-    TextModel,
+    # TextModel,
+    ImageObject,
+    # TextObject,
 )
 
 Image: ImageModel = ImageModel
-Text: TextModel = TextModel
+# Text: TextModel = TextModel
 
 samples: list = ext.samples
 styles: dict = ext.styles
@@ -77,9 +81,9 @@ def help(*args, **kwargs) -> str:
 
   Image.image(): return the image object
 
-  Text(save_file: str (OPTIONAL)): inialize the ai.Text
+  (deprecated) > Text(save_file: str (OPTIONAL)): inialize the ai.Text
 
-  Text.chat(prompt: str): chat with the ai and return the response
+  (deprecated) > Text.chat(prompt: str): chat with the ai and return the response
 
   """
         ""